Finsihed Code
Finished 1665 Robot Code

+++ src/main/java/frc/robot/Constants.java
@@ -1,12 +1,12 @@
 package frc.robot;
 public class Constants{
 
-    public static final double BALL_ONE = 4096*280.8;
-    public static final double BALL_TWO = 4096*540;
-    public static final double BALL_THREE = 4096*810;
-    public static final double HATCH_ONE = 4096*118.8;
-    public static final double HATCH_TWO = 4096*432;
-    public static final double HATCH_THREE = 4096*669.6;
-    public static final double GROUND = 0;
-    public static final double CARGO_STATION = 4096*378;
+    public static final int BALL_ONE = (int)(4096*280.8);
+    public static final int BALL_TWO = (int)(4096*540);
+    public static final int BALL_THREE = (int)(4096*810);
+    public static final int HATCH_ONE = (int)(4096*118.8);
+    public static final int HATCH_TWO = (int)(4096*432);
+    public static final int HATCH_THREE = (int)(4096*669.6);
+    public static final int GROUND = 0;
+    public static final int CARGO_STATION = (int)(4096*378);
 }
